+++ features/esbabbler/v5.md
@@ -58,7 +58,7 @@ Deferred in v3 ("Requires SFU"). Now unblocked.
 **Client components:**
 
 - [x] `pages/call/index.vue` — standalone call lobby with new-call and join-by-code/link actions
-- [x] `Content/Call/ControlBar.vue` — composed from single-purpose mute, audio settings, camera, video settings, deafen, screenshare, and leave controls
+- [x] `Content/Call/ControlBar.vue` — composed from single-purpose Google Meet-style grouped mic/settings, camera/settings, deafen, screenshare, and leave controls
 
 **Models:**
 
@@ -78,7 +78,7 @@ Deferred in v3 ("Requires SFU"). Now unblocked.
 
 - [x] `CallParticipant.ts` — `isCameraEnabled` field
 - [x] `Content/Call/CameraButton.vue` — camera toggle button
-- [x] `Content/Call/VideoSettingsButton.vue` — camera device selector
+- [x] `Content/Call/VideoSettingsButton.vue` — camera device selector and starter virtual backgrounds
 - [x] `Content/Call/ParticipantTile.vue` — renders camera tracks with avatar fallback, speaking ring, and mute badge
 - [x] `Content/Call/View.vue` — adaptive video grid for the fullscreen call surface
 - [x] `store/message/room/call.ts` — local/remote camera stream state
