+++ packages/app/app/components/Message/Content/Call/AudioControlGroup.vue
@@ -1,0 +1,6 @@
+<template>
+  <div rd-full bg-background flex items-center>
+    <MessageContentCallMuteButton />
+    <MessageContentCallAudioSettingsButton />
+  </div>
+</template>
