+++ packages/app/app/components/Message/Content/Call/CameraButton.vue
@@ -4,9 +4,6 @@ import { useCallStore } from "@/store/message/room/call";
 const callStore = useCallStore();
 const { isCameraEnabled } = storeToRefs(callStore);
 const { toggleCamera } = callStore;
-const onClick = async () => {
-  await toggleCamera();
-};
 </script>
 
 <template>
@@ -15,6 +12,6 @@ const onClick = async () => {
     :icon="isCameraEnabled ? 'mdi-video' : 'mdi-video-off'"
     :tooltip="isCameraEnabled ? 'Turn Camera Off' : 'Turn Camera On'"
     variant="plain"
-    @click="onClick()"
+    @click="toggleCamera()"
   />
 </template>
