+++ packages/app/app/components/Message/Content/Call/ControlBar.vue
@@ -1,10 +1,8 @@
 <template>
   <div pb-4 flex justify-center>
     <StyledCard px-4 py-2 rd-full flex gap-x-2 items-center>
-      <MessageContentCallMuteButton />
-      <MessageContentCallAudioSettingsButton />
-      <MessageContentCallCameraButton />
-      <MessageContentCallVideoSettingsButton />
+      <MessageContentCallAudioControlGroup />
+      <MessageContentCallVideoControlGroup />
       <MessageContentCallDeafenButton />
       <MessageContentCallScreenShareButton />
       <MessageContentCallLeaveButton />
