+++ packages/app/app/components/Message/Content/Call/MuteButton.vue
@@ -4,9 +4,6 @@ import { useCallStore } from "@/store/message/room/call";
 const callStore = useCallStore();
 const { isMuted } = storeToRefs(callStore);
 const { toggleMute } = callStore;
-const onClick = async () => {
-  await toggleMute();
-};
 </script>
 
 <template>
@@ -15,6 +12,6 @@ const onClick = async () => {
     :icon="isMuted ? 'mdi-microphone-off' : 'mdi-microphone'"
     :tooltip="isMuted ? 'Unmute' : 'Mute'"
     variant="plain"
-    @click="onClick()"
+    @click="toggleMute()"
   />
 </template>
