+++ packages/app/app/components/Message/Content/Call/ParticipantTile.vue
@@ -18,7 +18,6 @@ const displayName = computed(() => (isSelf ? `${participant.name} (You)` : parti
 <template>
   <div
     b-1
-    b-border
     rd-2
     b-solid
     bg-surface
@@ -30,8 +29,9 @@ const displayName = computed(() => (isSelf ? `${participant.name} (You)` : parti
     relative
     overflow-hidden
     elevation-3
+    :class="isSpeaking ? 'b-primary' : 'b-border'"
   >
-    <div v-if="isSpeaking" b-2 b-primary rd-2 b-solid pointer-events-none inset-0 absolute animate-pulse />
+    <div v-if="isSpeaking" b-2 b-primary rd-2 b-solid pointer-events-none z-1 inset-0 absolute animate-pulse />
     <video
       v-if="videoStream"
       autoplay
@@ -43,7 +43,7 @@ const displayName = computed(() => (isSelf ? `${participant.name} (You)` : parti
       :srcObject.prop="videoStream"
       :muted="isSelf"
     />
-    <div v-else bg-surface flex size-full items-center justify-center>
+    <div v-else bg-surface-opacity-80 flex size-full items-center justify-center>
       <StyledAvatar :image="participant.image" :name="participant.name" :avatar-props="{ size: '6rem' }" />
     </div>
     <div m-2 px-2 py-1 rd bg-surface-opacity-80 flex gap-x-2 items-center bottom-0 left-0 absolute>
