+++ packages/app/app/components/Message/Content/Call/VideoControlGroup.vue
@@ -1,0 +1,6 @@
+<template>
+  <div bg-background rd-full flex items-center>
+    <MessageContentCallCameraButton />
+    <MessageContentCallVideoSettingsButton />
+  </div>
+</template>
