+++ packages/app/app/services/message/room/call/CallVirtualBackgroundDefinitions.ts
@@ -1,0 +1,22 @@
+export const CallVirtualBackgroundDefinitions = [
+  {
+    imagePath: "",
+    title: "None",
+  },
+  {
+    imagePath: "/images/call/backgrounds/sunrise.svg",
+    title: "Sunrise",
+  },
+  {
+    imagePath: "/images/call/backgrounds/forest.svg",
+    title: "Forest",
+  },
+  {
+    imagePath: "/images/call/backgrounds/aurora.svg",
+    title: "Aurora",
+  },
+  {
+    imagePath: "/images/call/backgrounds/studio.svg",
+    title: "Studio",
+  },
+] as const;
